connect frontend and backend finished

+++ src/components/users/UserList.jsx
@@ -1,12 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import { useUserContext } from '../../contexts/UserContext';
 import UserModal from "./UserModal";
+import { getLocal } from "../utils/Storage.js";
 import { Container, Table, Button } from 'react-bootstrap';
 
 const UserList = () => {
   const { userApi } = useUserContext();
   const [openUserModal, setOpenUserModal] = useState(false);
-
+  const isAdmin = getLocal('isAdmin');
+console.log(isAdmin, 33);
   const openModal = () => {
     setOpenUserModal(true);
   }
@@ -23,7 +25,7 @@ const UserList = () => {
     <Container className="users rounded user-list">
       <UserModal show={openUserModal} onHide={closeModal} user={userApi.state.user} />
         <div>
-          <Button variant="primary" className="btn btn-sm" onClick={()=> {userApi.initUser(); openModal();}}><small>Create new user</small></Button>
+          { !isAdmin && <Button variant="primary" className="btn btn-sm" onClick={()=> {userApi.initUser(); openModal();}}><small>Create new user</small></Button>}
           <h6 className="text-right"><small>Total users is {userApi.state.counter}</small></h6>
         </div>
         <Table>
@@ -36,12 +38,12 @@ const UserList = () => {
               <th>Email</th>
               <th>CreatedAt</th>
               <th>UpdatedAt</th>
-              <th>Operation</th>
+              { !isAdmin && <th>Operation</th>}
             </tr>
           </thead>
           <tbody>
             {userApi.state.listUsers.map((user, index) => (
-              <tr key={index}>
+              <tr key={index} id={user.id}>
                 <td>{user.username}</td>
                 <td>{user.first_name}</td>
                 <td>{user.last_name}</td>
@@ -50,8 +52,8 @@ const UserList = () => {
                 <td>{user.created_at}</td>
                 <td>{user.updated_at}</td>
                 <td>
-                  <Button variant="primary" className="btn btn-sm" onClick={() => {userApi.editUser(user); openModal();}}>Update</Button>{' '}
-                  <Button variant="danger" className="btn btn-sm" onClick={() => userApi.deleteUser(user.id)}>Delete</Button>
+                  { !isAdmin && <Button variant="primary" className="btn btn-sm" onClick={() => {userApi.editUser(user); openModal();}}>Update</Button> }
+                  { !isAdmin && <Button variant="danger" className="btn btn-sm" onClick={() => {userApi.deleteUser(user.id); document.getElementById(user.id).style.display="none";}}>Delete</Button>}
                 </td>
               </tr>
             ))}
